# Conflicts:
#	src/main/java/edu/mum/mumsched/model/Section.java
#	src/main/java/edu/mum/mumsched/model/Student.java

+++ src/main/java/edu/mum/mumsched/controller/CourseController.java
@@ -1,7 +1,14 @@
 package edu.mum.mumsched.controller;
 
+import edu.mum.mumsched.SectionRegistrationSubsystem.ISectionRegistrationSubsystem;
 import edu.mum.mumsched.model.Course;
+import edu.mum.mumsched.model.Faculty;
+import edu.mum.mumsched.model.Section;
+import edu.mum.mumsched.model.Specialization;
 import edu.mum.mumsched.service.ICourseService;
+import edu.mum.mumsched.service.IFacultyService;
+import edu.mum.mumsched.service.IStudentService;
+import edu.mum.mumsched.util.SpecializationUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -10,14 +17,25 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
+import java.util.List;
 
 @Controller
 public class CourseController {
 
     @Autowired
     private ICourseService iCourseService;
+
+    @Autowired
+    private IFacultyService iFacultyService;
+
+    @Autowired
+    private ISectionRegistrationSubsystem sectionRegistrationFacade;
+
+    @Autowired
+    private IStudentService iStudentService;
 
     @GetMapping("/allCourse")
     public String findAll(Model model){
@@ -29,6 +47,8 @@ public class CourseController {
     public String addCourseForm(Model model){
         model.addAttribute("course", new Course());
         model.addAttribute("prerequisiteList", iCourseService.findAll());
+        model.addAttribute("facultyList", iFacultyService.findAll());
+        model.addAttribute("specializationList", SpecializationUtil.getSpecializations());
         return "course/form";
     }
 
@@ -36,19 +56,49 @@ public class CourseController {
     public String updateCourseForm(@PathVariable("id") Integer id, Model model){
         model.addAttribute("course", iCourseService.findById(id));
         model.addAttribute("prerequisiteList", iCourseService.findAllExcept(id));
+        model.addAttribute("facultyList", iFacultyService.findAll());
+        model.addAttribute("specializationList", SpecializationUtil.getSpecializations());
         return "course/form";
     }
 
     @PostMapping("/saveCourse")
-    public String saveCourse(@ModelAttribute @Valid Course course, BindingResult bindingResult){
+    public String saveCourse(@ModelAttribute @Valid Course course, Model model,
+                             BindingResult bindingResult, RedirectAttributes redirectAttributes){
         bindingResult.hasErrors();
+        List<Course> duplicates = iCourseService.findDuplicates(course.getName(), course.getId());
+        if (duplicates.size() > 0) {
+            redirectAttributes.addFlashAttribute("messageError", "The course name cannot be duplicated.");
+            if (course.getId() > 0) {
+                return "redirect:/updateCourse/" + course.getId();
+            } else {
+                return "redirect:/newCourse";
+            }
+        }
         iCourseService.save(course);
         return "redirect:/allCourse";
     }
 
     @GetMapping("/deleteCourse/{id}")
-    public String deleteCourseForm(@PathVariable("id") Integer id){
-        iCourseService.delete(id);
+    public String deleteCourseForm(@PathVariable("id") Integer id, RedirectAttributes redirectAttributes){
+        Course course = iCourseService.findById(id);
+        List<Course> prereqs = iCourseService.findAllByPrerequisitesContains(course);
+        List<Faculty> faculties = iFacultyService.findAllByPreferredCoursesContains(course);
+        List<Section> sections = sectionRegistrationFacade.findByCourse(course);
+        if(!prereqs.isEmpty() || !faculties.isEmpty() || !sections.isEmpty()) {
+            String messages = "The course was not deleted!";
+            if(!prereqs.isEmpty()) {
+                messages += " Course is used for prerequisite.";
+            }
+            if(!faculties.isEmpty()) {
+                messages += " Course is used for Faculty's preferred course.";
+            }
+            if(!sections.isEmpty()) {
+                messages += " Course is used for Section.";
+            }
+            redirectAttributes.addFlashAttribute("messageError", messages);
+        } else {
+            iCourseService.delete(id);
+        }
         return "redirect:/allCourse";
     }
